requirements.txt
 python-dotenv
customtkinter
SpeechRecognition
edge-tts
pygame
pyautogui
pyperclip
psutil
google-generative-ai
groq
openai
pyaudio
pillow
opencv-contrib-python
deepface
tf-keras
chromadb
sentence-transformers
PyPDF2
playwright
google-auth-oauthlib
